% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdplotest.R
\name{fwdplotest}
\alias{fwdplotest}
\title{Forward plots of summary estimates.}
\usage{
fwdplotest(x)
}
\arguments{
\item{x}{object of class NMAoutlier (mandatory).}
}
\value{
forward plots of summary estimates.
}
\description{
Forward plots of summary estimate with 95 percent confidence interval for each treatment.
}
\details{
Plot of summary estimates and their confidence intervals for each iteration of search.
Vertical axis provides iterations of search. Horizontal axis provides summary estimate of a treatment.
}
\examples{
\dontrun{
# We use the smoking cessation dataset from netmeta package

data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package

library(netmeta)
p1 <- pairwise(list(treat1, treat2, treat3),
              list(event1, event2, event3),
              list(n1, n2, n3),
              data=smokingcessation,
              sm="OR")
p1


# forward search algorithm
FSresult <- NMAoutlier(p1, small.values = "bad")


# forward plot for summary estimates for each treatment
# and their confidence intervals
fwdplotest(FSresult)

}

}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
