% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indices.R
\name{Indices}
\alias{Indices}
\title{Indices}
\usage{
Indices(studlab, set)
}
\arguments{
\item{studlab}{Study labels (important when multi arm studies are included).}

\item{set}{set of studies}
}
\value{
indices for a set of studies
}
\description{
Finds the indices of a set of studies (this is helpful to find the indices for the case of a multi-arm trial).
}
