% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.NMAoutlier.R
\name{measures.NMAoutlier}
\alias{measures.NMAoutlier}
\title{Outlier and influential detection measures in network meta-analysis.}
\usage{
measures.NMAoutlier(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  data = NULL,
  sm,
  reference = "",
  measure = "influential"
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{data}{A data frame containing the study information.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group.}

\item{measure}{Outlier and influential detection measures (default: "influential")
and measures considered study deletion (measure = "deletion").}
}
\value{
An object of class \code{NMAoutlier_measures}; a list containing the
following components:
   \item{dat}{Matrix containing the data \code{"TE"}, \code{"seTE"}, \code{"studlab"}, \code{"treat1"}, \code{"treat2"} as defined above.}
   \item{eraw}{Raw residual for each study included in the network.}
   \item{estand}{Standardized residual for each study included in the network.}
   \item{estud}{Studentized residual for each study included in the network.}
   \item{Mah}{Mahalanobis distance for each study included in the network.}
   \item{leverage}{Leverage for each study included in the network.}
   \item{call}{Function call}
}
\description{
Employs the computation of several measures for detection of outlying
studies (studies with extreme results) and influential studies fitted in network
meta-analysis model from graph-theory. Calculation of these measures for each study
included in the network can detect evidence of outliers. It can also be used to detect
studies that are potential sources for heterogeneity and
inconsistency.

Statistical outlier and influential measures are:
\itemize{
\item outlying and influential measures for each study (raw residuals, standardized residuals,
 studentized residuals, Mahalanobis distance, leverage).
}

A description of the several outlier detection measures in the context of network meta-analysis
can be found in Petropoulou et al. (2020).
}
\details{
Description of several outlier and influential measures are calculated for
the network meta-analysis model from graph theory (R端cker, 2012)
fitted with (\code{netmeta} function) of R package \bold{netmeta} (R端cker et al., 2015).
The researcher can choose the reference treatment \code{reference} fitted in NMA model.

An overview of the several outlier detection measures is described in Petropoulou et al. 2020.

Let \emph{n} be the number of treatments in a network and let
\emph{m} be the number of pairwise treatment comparisons.  If there
are only two-arm studies, \emph{m} is the number of studies. Let
\code{TE} and \code{seTE} be the vectors of observed effects and their standard
errors.  Comparisons belonging to multi-arm studies are identified
by identical study labels (argument \code{studlab}). It is
therefore important to use identical study labels for all
comparisons belonging to the same multi-arm study.

The function calculates several outlier detection measures for each study.
The statistical measures calculated are:
Raw residuals, Standardized residuals, Studentized residuals, Mahalanobis distance
and leverage for each study.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")


#' # outlier and influential detection measures for studies 9, 10, 11, 12
meas <- measures.NMAoutlier(p1)

# Standardized residual for each study included in the network
meas$estand


# deletion measures for studies 9, 10, 11, 12.
delete <- measures.NMAoutlier(p1, measure = "deletion")

# Standardized deleted residual for studies 9, 10, 11, 12.
delete$estand.deleted


\dontrun{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm = "OR")

# outlier and influential detection measures for each study in the network
meas <- measures.NMAoutlier(p1, measure = "influential")

# Mahalanobis distance for each study included in the network
meas$Mah
}

}
\references{
R端cker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24

R端cker G, Schwarzer G (2015):
Ranking treatments in frequentist network meta-analysis works
without resampling methods.
\emph{BMC Medical Research Methodology},
\bold{15}, 58

Petropoulou M (2020):
Exploring methodological challenges in network meta-analysis models and
developing methodology for outlier detection.
\emph{PhD dissertation}
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
