% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{mult}
\usage{
mult(studlab, ind.bs, standres, model)
}
\arguments{
\item{studlab}{Study labels (important when multi arm studies are included).}

\item{ind.bs}{indices of basic set.}

\item{standres}{standardized residuals of each pairwise comparison.}

\item{model}{model assigned from netmeta}
}
\value{
standardized residuals and arithmetic mean of standardized residuals in case of multi-arm trials.
}
\description{
Computes the arithmetic mean of standardized residuals in case of multi-arm trials.
}
\keyword{internal}
