% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.R
\name{Multi}
\alias{Multi}
\title{Residuals and log-likelihood contributions in multi-arm trials}
\usage{
Multi(st.lab, ei, wr)
}
\arguments{
\item{st.lab}{studylab of set of studies}

\item{ei}{standardized residuals for each comparison from a set of studies}

\item{wr}{weigths of random effects model}
}
\value{
the arithmetic mean of standardized and log-likelihood contributions.
}
\description{
Computes the arithmetic mean of standardized and log-likelihood contributions in the case of multi-arm trials.
}
\keyword{internal}
