% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmet.R
\name{netmet}
\alias{netmet}
\title{Network meta-analysis for a set of studies}
\usage{
netmet(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  ind.bs,
  reference,
  small.values,
  names.treat,
  ...
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{ind.bs}{indices of studies in basic set}

\item{reference}{Reference treatment group}

\item{small.values}{small.values for P-score. A character string
indicating if small values are considered bad ("bad") or good
("good").}

\item{names.treat}{names of treatments}

\item{\dots}{Additional arguments passed on to
\code{\link{netmeta}}.}
}
\value{
results and statistics from network meta-analysis.
}
\description{
Conduct network meta-analysis (Rücker model) for a set of studies.
}
\keyword{internal}
