% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.R
\name{nma}
\alias{nma}
\title{Network meta-analysis for the whole dataset.}
\usage{
nma(TE, seTE, treat1, treat2, studlab, reference, names.treat, ...)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{reference}{Reference treatment group}

\item{names.treat}{names of treatments}
}
\value{
Within-study standard error, design matrix, summary estimate,
heterogeneity from network meta-analysis.
}
\description{
Conduct network meta-analysis (Rücker model) for the whole dataset.
}
\keyword{internal}
