% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_REML.R
\name{NMA_REML}
\alias{NMA_REML}
\title{Restricted Maximum log-likelihood (REML) function for NMA model.}
\usage{
NMA_REML(
  x,
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  Xdesign,
  tr1,
  tr2,
  nt,
  observations
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{Xdesign}{design matrix.}

\item{tr1}{positions for the first treatment.}

\item{tr2}{positions for the second treatment.}

\item{nt}{number of treatments.}

\item{observations}{(optional) (default: NA) The study (or studies) to fit by shift its (their) variance.
the default value is NA and shift variance model is fitted for each study.}
}
\value{
REML log-likelihood function.
}
\description{
Conduct REML log-likelihood function for NMA model.
}
\keyword{internal}
