% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMA_RVSOM.R
\name{NMA_RVSOM}
\alias{NMA_RVSOM}
\title{Random shift variance network meta-analysis (RSV NMA) model.}
\usage{
NMA_RVSOM(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  reference,
  small.values,
  observations
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{reference}{Reference treatment group.}

\item{small.values}{A character string indicating if small values
are considered beneficial (option:"good") or harmfull
(option:"bad") on outcome.  This is requirement for p-scores
computation. The default value is considered benefial outcome
("good").}

\item{observations}{(optional) (default: NA) The study (or studies) to fit by shift its (their) variance.
the default value is NA and shift variance model is fitted for each study.}
}
\value{
A list containing the following components:
}
\description{
Conduct RSV NMA model.
}
\keyword{internal}
