% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAoutlier.measures.R
\name{NMAoutlier.measures}
\alias{NMAoutlier.measures}
\title{Outlier and influential detection measures in network meta-analysis.}
\usage{
NMAoutlier.measures(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  data = NULL,
  sm,
  reference = "",
  measure = "simple"
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{data}{A data frame containing the study information.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group.}

\item{measure}{Outlier and influential detection measures. Simple measures (default: "simple")
and measures considered study deletion (measure = "deletion").}
}
\value{
An object of class \code{NMAoutlier.measures};
when choosing simple detection measures, a list containing the following components:
   \item{dat}{Matrix containing the data \code{"TE"}, \code{"seTE"}, \code{"studlab"}, \code{"treat1"}, \code{"treat2"} as defined above.}
   \item{eraw}{Raw residual for each study included in the network.}
   \item{estand}{Standardized residual for each study included in the network.}
   \item{estud}{Studentized residual for each study included in the network.}
   \item{Mah}{Mahalanobis distance for each study included in the network.}
   \item{leverage}{Leverage for each study included in the network.}
   \item{measure}{type of measure used.}
   \item{call}{Function call}

when choosing detection measures considered deletion, a list containing the following components:
following components:
   \item{dat}{Matrix containing the data \code{"TE"}, \code{"seTE"}, \code{"studlab"}, \code{"treat1"}, \code{"treat2"} as defined above.}
   \item{eraw.deleted}{Raw deleted residual for each study included in the network.}
   \item{estand.deleted}{Standardized deleted residual for each study included in the network.}
   \item{estud.deleted }{Studentized deleted residual for each study included in the network.}
   \item{Cooks.distance}{Cook distance between the treatment estimates for study j and treatment estimates when study j is removed}
   \item{Covratio}{Ratio of determinants of variance-covariance matrix of treatment estimates for study j to treatment estimates when study j is removed.}
   \item{w.leaveoneout}{Weight leave one out.}
   \item{H.leaveoneout}{leverage leave one out.}
   \item{heterog.leaveoneout}{heterogeneity estimator leave one out.}
   \item{Rheterogeneity}{R statistic for heterogeneity.}
   \item{Restimates}{R statistis for estimates.}
   \item{RQtotal}{R statistic for Q (\code{Qtotal}).}
   \item{RQhet}{R statistic for  heterogeneity Q (\code{Qhet}).}
   \item{RQinc}{R statistic for Qinconsistency (\code{Qinc}).}
   \item{DFbetas}{DFbetas.}
   \item{measure}{type of measure used.}
   \item{call}{Function call}
}
\description{
Employs the computation of several (simple or deletion) measures for detection of outlying
studies (studies with extreme results) and influential studies fitted in network
meta-analysis model from graph-theory.

Statistical outlier and influential measures are:
\itemize{
\item outlying and influential measures for each study (raw residuals, standardized residuals,
 studentized residuals, Mahalanobis distance, leverage).
}
}
\details{
Several outlier and influential measures (simple or deletion) can be calculated.
The network meta-analysis model from graph theory (R端cker, 2012)
fitted with (\code{netmeta} function) of R package \bold{netmeta} (R端cker et al., 2015).
The researcher can choose the reference treatment \code{reference} fitted in NMA model.

A description of the several outlier detection measures in the context of network meta-analysis
can be found in Petropoulou (2020).

Let \emph{n} be the number of treatments in a network and let
\emph{m} be the number of pairwise treatment comparisons.  If there
are only two-arm studies, \emph{m} is the number of studies. Let
\code{TE} and \code{seTE} be the vectors of observed effects and their standard
errors.  Comparisons belonging to multi-arm studies are identified
by identical study labels (argument \code{studlab}). It is
therefore important to use identical study labels for all
comparisons belonging to the same multi-arm study.

The function calculates several outlier detection measures for each study.
The statistical measures calculated are:
For simple outlier and influential measures (measure = "simple"):
Raw residuals, Standardized residuals, Studentized residuals, Mahalanobis distance
and leverage for each study.
For deletion outlier and influential measures (measure = "deletion"):
Standardized deleted residual;
Studentized deleted residual; Cook distance between the treatment estimates for study j
and treatment estimates when study j is removed;
Ratio of determinants of variance-covariance matrix of treatment estimates for study j to treatment estimates when study j is removed;
Weight leave one out;leverage leave one out; heterogeneity estimator leave one out;
R statistic for heterogeneity;  R statistis for estimates; R statistic for Q (\code{Qtotal}),  R statistic for  heterogeneity Q
(\code{Qhet}), R statistic for Qinconsistency (\code{Qinc}), DFbetas.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")


#' # outlier and influential detection measures for studies 9, 10, 11, 12
meas <- NMAoutlier.measures(p1)

# Standardized residual for each study included in the network
meas$estand


# deletion measures for studies 9, 10, 11, 12.
delete <- NMAoutlier.measures(p1, measure = "deletion")

# Standardized deleted residual for studies 9, 10, 11, 12.
delete$estand.deleted


\dontrun{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm = "OR")

# outlier and influential detection measures for each study in the network
meas <- NMAoutlier.measures(p1, measure = "simple")

# Mahalanobis distance for each study included in the network
meas$Mah
}

}
\references{
R端cker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24

R端cker G, Schwarzer G (2015):
Ranking treatments in frequentist network meta-analysis works
without resampling methods.
\emph{BMC Medical Research Methodology},
\bold{15}, 58

Petropoulou M (2020):
Exploring methodological challenges in network meta-analysis models and
developing methodology for outlier detection.
\emph{PhD dissertation}
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
