% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAoutlier.R
\name{NMAoutlier}
\alias{NMAoutlier}
\title{Forward Search Algorithm in Network Meta-Analysis (NMA).}
\usage{
NMAoutlier(TE, seTE, treat1, treat2, studlab, data = NULL, crit1 = "R",
  crit2 = "R", studies = NULL, P = 100, sm, reference = "",
  small.values = "good")
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between first and second treatment (e.g. log odds ratio, mean difference, or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are included).}

\item{data}{A data frame containing the study information.}

\item{crit1}{A character string indicating the criterion to be used for selecting the initial subset,  this criterion could be the minimum of median absolute residuals ("R") or the maximum of median absolute likelihood contributions ("L"). Default value is "R".}

\item{crit2}{A character string indicating the criterion to be used for selecting the study from non-basic set to basic set, this criterion could be the minimum of absolute residuals ("R") or the maximum of absolute likelihood contributions ("L"). Default value is "R".}

\item{studies}{An optional vector specifying the number of the initial subset of studies. The default value is the maximum of the number of treatments and the rounded number of 20 percent of the total number of studies.}

\item{P}{An optional vector specifying the number of samples for the choice of the initial subset. Default value is 100.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group}

\item{small.values}{A character string indicating if small values are considered beneficial (option:"good") or harmfull (option:"bad") on outcome. This is requirement for p-scores computations.}
}
\value{
Provide the data, give the series of studies that entered in FS algorithm and the number of iteration of FS, outlier and influential case diagnostics  measures (standardized and studentized residuals, Cook distance, Ratio of variances), heterogeneity and inconsistency measures (heterogeneity, Q statistics (Q for heterogeneity, Q for inconsistency, Q total), Node-splitting (difference of direct and indirect evidence and its confidence interval)), ranking measures (P-score).
}
\description{
Conduct Forward Search (FS) Algorithm in Network Meta-Analysis (NMA).
}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
