% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAoutlier.rsv.R
\name{NMAoutlier.rsv}
\alias{NMAoutlier.rsv}
\title{Random Shift Variance Network Meta-analysis (RSV NMA) model.}
\usage{
NMAoutlier.rsv(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  data = NULL,
  sm,
  reference = "",
  small.values = "good",
  study = NULL,
  n_cores = NULL
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{data}{A data frame containing the study information.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group.}

\item{small.values}{A character string indicating if small values
are considered beneficial (option:"good") or harmfull
(option:"bad") on outcome.  This is requirement for p-scores
computation. The default value is considered benefial outcome
("good").}

\item{study}{The study (or studies) to fit by shift its (their) variance.
By default, the standard random effects model is fitted, value NULL.}

\item{n_cores}{The number of cores that the process is running
using the parallel (default: NULL, the process is running
using all the available cores)}
}
\value{

}
\description{
This function employs the Random Shift Variance Network Meta-analysis model to detect outliers
and influential studies fitted in network meta-analysis model from graph-theory.
This is an outlying diagnostic tool to detect outliers and studies that are potential
sources for heterogeneity and inconsistency.

Monitoring measures during the search are:
\itemize{
\item outlier detection measures (standardized residuals, leverage);
\item ranking measures (P-scores);
\item heterogeneity and inconsistency measures (Q statistics for
  overall heterogeneity / inconsistency, inconsistency by
  design-by-treatment interaction model);
\item over-dispersion parameter (shift variance estimator);
\item Likelihood Ratio Test (LRT)
}

A description of the methodology can be found in Petropoulou et al. (2020).
}
\details{
Random Shift Variance Network Meta-Analysis (RSV NMA) model is in described in Petropoulou (2020).

The RSV NMA model is fitted by shifting the study variance for each included study.
RSV NMA model is an extension of the standard network meta-analysis model
from graph theory (Rücker, 2012) (\code{netmeta} function) fitted with
R package \bold{netmeta} (Rücker et al., 2015). RSV NMA model is based on the REstricted Maximum Likelihood (REML)
estimation method for heterogeneity variance. The variance estimator of RSV NMA model is
decomposed by the heterogeneity variance and the over-dispersion parameter.
Monitoring by shifting each study variance is helpful to identify outlying and/or influential studies.


Monitoring statistical measures for each fit can be:

\bold{Likelihood statistics:}
Heterogeneity estimation method is conducted under
the Restricted Maximum likelihood estimator.
The variance estimator of shift variance model.
Likelihood statistics offered from calculation:
the twice of maximum log-likelihood;
the convergence diagnostic and the Likelihood Ratio test (LRT) test.

\bold{Outlier and influential case diagnostics measures:}
Standardized residuals (arithmetic mean in case of multi-arm
studies); leverage of hat matrix

\bold{Ranking measures:}
P-scores for ranking of treatments (Rücker G & Schwarzer G (2015))
for each basic set with implementation of (\code{netrank} function)
from R package \bold{netmeta}.

\bold{Heterogeneity and inconsistency measures:}
Overall heterogeneity / inconsistency Q statistic (\code{Q}) This
is the design-based decomposition of Cochran Q as provided by Krahn
et al.(2013); Overall heterogeneity Q statistic (\code{Q});
Between-designs Q statistic (\code{Q}), based on a random effects
model with square-root of between-study variance estimated embedded
in a full design-by-treatment interaction model. Implementation
with (\code{decomp.design} function) from R package \bold{netmeta};
Z-values (Dias et al., 2010; Konig et al., 2013) for comparison
between direct and indirect evidence in
each iteration of forward search algorithm. By monitoring
difference of direct and indirect evidence, potential sources of
consistency can be detected with the implementation of
(\code{netsplit} function) from R package \bold{netmeta} for each
iteration of the search.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm="OR")

# RSV NMA model implementation for study 1
#
RVSOMres <- NMAoutlier.rsv(p1,  small.values = "bad", study = c(1) , n_cores = 2)
#

\dontrun{
# Random Shift Variance (RSV NMA) model
#
RVSOMresult <- NMAoutlier.rsv(p1, small.values = "bad")

# Random shift variance estimator
RVSOMresult$over-disp
#
# Plot of Likelihood Ratio Test
RVSOMresult$LRT
#
# Summary treatment estimates and their 95\% confidence intervals
# by downweighting each study
RVSOMresult$b
RVSOMresult$l
RVSOMresult$u
}

}
\references{
Rücker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24

Rücker G, Schwarzer G (2015):
Ranking treatments in frequentist network meta-analysis works
without resampling methods.
\emph{BMC Medical Research Methodology},
\bold{15}, 58

Petropoulou M (2020):
Exploring methodological challenges in network meta-analysis models and
developing methodology for outlier detection.
\emph{PhD dissertation}.

Krahn U, Binder H, König J (2013):
A graphical tool for locating inconsistency in network meta-analyses.
\emph{BMC Medical Research Methodology},
\bold{13}, 35

Dias S, Welton NJ, Caldwell DM, Ades AE (2010):
Checking consistency in mixed treatment comparison meta-analysis.
\emph{Statistics in Medicine},
\bold{29}, 932--44

König J, Krahn U, Binder H (2013):
Visualizing the flow of evidence in network meta-analysis and
characterizing mixed treatment comparisons.
\emph{Statistics in Medicine},
\bold{32}, 5414--29
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
