% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAoutlier_deletion_measures.R
\name{NMAoutlier_deletion_measures}
\alias{NMAoutlier_deletion_measures}
\title{Outlier and influential detection measures in network meta-analysis considered deletion of study.}
\usage{
NMAoutlier_deletion_measures(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  data = NULL,
  sm,
  reference = ""
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{data}{A data frame containing the study information.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group.}
}
\value{
An object of class \code{NMAoutlier_measures}; a list containing the
following components:
   \item{dat}{Matrix containing the data \code{"TE"}, \code{"seTE"}, \code{"studlab"}, \code{"treat1"}, \code{"treat2"} as defined above.}
   \item{eraw.deleted}{Raw deleted residual for each study.}
   \item{estand.deleted}{Standardized deleted residual for each study.}
   \item{estud.deleted}{Studentized deleted residual for each study.}
   \item{Cooks.distance}{Cook's distance.}
   \item{Covratio}{ratio of determinants of variances of variance-covariance matrix of treatment estimates.}
   \item{w.leaveoneout}{weight leave one out.}
   \item{H.leaveoneout}{leverage leave one out.}
   \item{heterog.leaveoneout}{heterogeneity estimator leave one out.}
   \item{Rheterogeneity}{R statistic for heterogeneity.}
   \item{RQtotal}{R statistic for Qtotal.}
   \item{RQhet}{R statistic for Qheterogeneity.}
   \item{RQinc}{R statistic for Qinconsistency.}
   \item{DFbetas}{DFbetas.}
   \item{call}{Function call}
}
\description{
Employs the computation of several measures for detection of outlying
studies (studies with extreme results) and influential studies considering study deletion fitted in network
meta-analysis model from graph-theory. Calculation of these measures for each study
included in the network can detect evidence of outliers. It can also be used to detect
studies that are potential sources for heterogeneity and
inconsistency.

Statistical outlier and influential deletion measures are:
\itemize{
\item outlying and influential measures for each study are:
Standardized deleted residual;Studentized deleted residual; Cook
distance between the treatment estimates for study j and treatment estimates when study j is removed;
ratio of determinants of variances of variance-covariance matrix
of treatment estimates for study j to treatment estimates when study j is removed;
weight leave one out; leverage leave one out; heterogeneity estimator leave one out;
R statistic for heterogeneity;  R statistic for Q (\code{Qtotal}),  R statistic for  heterogeneity Q
(\code{Qhet}), R statistic for Qinconsistency (\code{Qinc}), DFbetas.
}

A description of the several outlier detection measures considering study deletion in the context of network meta-analysis
can be found in Petropoulou et al. (2019).
}
\details{
Description of several outlier and influential measures are calculated for
the network meta-analysis model from graph theory (R端cker, 2012)
fitted with (\code{netmeta} function) of R package \bold{netmeta} (R端cker et al., 2015).
The researcher can choose the reference treatment \code{reference} fitted in NMA model.

An overview of the several outlier detection measures is described in Petropoulou et al. 2019.

The function calculates several outlier detection measures for each study.
The statistical measures calculated such as are raw, standardized and studentized deleted residuals,
Cook distance, COVRATIO, weight leave one out etc.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")


#outlier and influential detection measures for studies 9, 10, 11, 12.
deletion <- NMAoutlier_deletion_measures(p1)

# Standardized deleted residual for studies 9, 10, 11, 12.
deletion$estand.deleted


\dontrun{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm = "OR")

# outlier and influential detection measures for each study in the network
deletion <- NMAoutlier_deletion_measures(p1)

# Cook's distance between the treatment estimates for study j
# and treatment estimates when study j is removed.
deletion$Cooks.distance


}

}
\references{
R端cker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24

R端cker G, Schwarzer G (2015):
Ranking treatments in frequentist network meta-analysis works
without resampling methods.
\emph{BMC Medical Research Methodology},
\bold{15}, 58

Petropoulou M (2019):
Outlier detection measures in network meta-analysis.
\emph{Manuscript}
}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
