% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NMAoutlier.R
\name{plot.NMAoutlier}
\alias{plot.NMAoutlier}
\title{Plot(s) to monitor selected outlier and influential measure(s) for each study.}
\usage{
\method{plot}{NMAoutlier}(object, stat, measure = "influential")
}
\arguments{
\item{object}{an object of class measures.NMAoutlier (mandatory).}

\item{stat}{selected statistical outlier and influential detection measure
(mandatory),
For oultier and influential measures available choices are: ("estand"/ "estud"/ "mah"/ "leverage").
For  oultier and influential deletion measures available choices are: (eraw.deleted", "estand.deleted", "estud.deleted",
"leverage.leaveoneout", "weight.leaveoneout", "heterog.leaveoneout", "covratio", "cook",
"rheterogeneity", "restimates", "rqhet", "rqinc", "rqtotal", "dfbetas")

 @param measure outlier and influential detection measures (default "influential")
 and measures considered study deletion (measure = "deletion").}
}
\description{
This function generates plot(s) of the selected outlier detection measure(s)
for each study included in the network. The function creates a plot of the
selected detection statistic for each study in the network.
Candidate statistics to be monitored can be Standardized residual;
Studentized residual; Mahalanobis distance and
leverage for each study.

#' This function also generates plot(s) of the selected outlier detection measure(s)
after the deletion of a study. The function creates a plot of the
selected detection statistic considering a deletion of each study in the network.
Candidate statistics to be monitored can be Raw deleted residual;
Standardized deleted residual;Studentized deleted residual; Cook
distance between the treatment estimates for study j and treatment estimates when study j is removed;
ratio of determinants of variances of variance-covariance matrix
of treatment estimates for study j to treatment estimates when study j is removed;
weight leave one out;leverage leave one out; heterogeneity estimator leave one out;
R statistic for heterogeneity;  R statistic for Q (\code{Qtotal}),  R statistic for  heterogeneity Q
(\code{Qhet}), R statistic for Qinconsistency (\code{Qinc}), DFbetas.
}
\details{
Plot of outlier and influential detection and deletion measures for each study included in the network.
Vertical axis provides the studies in the network (or the study deleted for deletion measures). Horizontal axis
provides a monitoring outlier and influential detection measure.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")

# outlier and influential detection measures for each study in the network
measures <- measures.NMAoutlier(p1)

# plot of Standardized residuals for each study
plot.NMAoutlier(measures, "estand")

# plot of Mahalanobis distance values for each study
plot.NMAoutlier(measures, "mah")

# plot of leverage values for each study
plot.NMAoutlier(measures, "leverage")

# outlier detection measures considered deletion each time of an included study
deletion <- measures.NMAoutlier(p1, measure = "deletion")

# plot for R statistic for heterogeneity estimator
plot.NMAoutlier(deletion, "rheterogeneity", measure = "deletion")

# plot for R statistic for Qinconsistency
plot.NMAoutlier(deletion, "rqinc", measure = "deletion")

# plot of COVRATIO values when considering deletion for each study
plot.NMAoutlier(deletion, "covratio", measure = "deletion")


}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{hplot}
