% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMAoutlier_deletion_measures.R
\name{plot_NMAoutlier_deletion_measures}
\alias{plot_NMAoutlier_deletion_measures}
\title{Plot(s) to monitor selected outlier statistical measure(s) after considered a deletion of a study.}
\usage{
plot_NMAoutlier_deletion_measures(object, stat)
}
\arguments{
\item{object}{an object of class NMAoutlier_deletion_measures  (mandatory).}

\item{stat}{selected statistical outlier and influential detection measure
considered deletion (mandatory), available choices are: ("eraw.deleted", "estand.deleted",
"estud.deleted", "heterog.leaveoneout", "covratio", "cook", "rheterogeneity",
"rqhet", "rqinc", "rqtotal").}
}
\description{
This function generates plot(s) of the selected outlier detection measure(s)
after the deletion of a study. The function creates a plot of the
selected detection statistic considering a deletion of each study in the network.
Candidate statistics to be monitored can be Raw deleted residual;
Standardized deleted residual;Studentized deleted residual; Cook
distance between the treatment estimates for study j and treatment estimates when study j is removed;
ratio of determinants of variances of variance-covariance matrix
of treatment estimates for study j to treatment estimates when study j is removed;
weight leave one out;leverage leave one out; heterogeneity estimator leave one out;
R statistic for heterogeneity;  R statistic for Q (\code{Qtotal}),  R statistic for  heterogeneity Q
(\code{Qhet}), R statistic for Qinconsistency (\code{Qinc}), DFbetas.
}
\details{
Plot of outlier and influential detection measures considering deletion of each study in the network.
Vertical axis provides the study deleted. Horizontal axis
provides a monitoring outlier and influential detection measure considered deletion.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")

# outlier detection measures considered deletion each time of an included study
deletion <- NMAoutlier_deletion_measures(p1)

# plot for R statistic for heterogeneity estimator
plot_NMAoutlier_deletion_measures(deletion, "rheterogeneity")

# plot for R statistic for Qinconsistency
plot_NMAoutlier_deletion_measures(deletion, "rqinc")

# plot of COVRATIO values when considering deletion for each study
plot_NMAoutlier_deletion_measures(deletion, "covratio")


}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
\keyword{hplot}
