% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMAoutlier_measures.R
\name{plot_NMAoutlier_measures}
\alias{plot_NMAoutlier_measures}
\title{Plot(s) to monitor selected outlier and influential  statistical measure(s) for each study.}
\usage{
plot_NMAoutlier_measures(object, stat)
}
\arguments{
\item{object}{an object of class NMAoutlier_measures (mandatory).}

\item{stat}{selected statistical outlier and influential detection measure
(mandatory), available choices are: ("estand"/ "estud"/ "mah"/ "leverage").}
}
\description{
This function generates plot(s) of the selected outlier detection measure(s)
for each study included in the network. The function creates a plot of the
selected detection statistic for each study in the network.
Candidate statistics to be monitored can be Standardized residual;
Studentized residual; Mahalanobis distance and
leverage for each study.
}
\details{
Plot of outlier and influential detection measures for each study included in the network.
Vertical axis provides the studies in the network. Horizontal axis
provides a monitoring outlier and influential detection measure.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)

study912 <- subset(smokingcessation, id \%in\% 9:12)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = study912,
                        sm = "OR")

# outlier and influential detection measures for each study in the network
measures <- NMAoutlier_measures(p1)

# plot of Standardized residuals for each study
plot_NMAoutlier_measures(measures, "estand")

# plot of Mahalanobis distance values for each study
plot_NMAoutlier_measures(measures, "mah")

# plot of leverage values for each study
plot_NMAoutlier_measures(measures, "leverage")


}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
\keyword{hplot}
