% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotesthelper.R
\name{plotesthelper}
\alias{plotesthelper}
\title{Helper function to create the plots of summary estimates and their intervals for FS and RSV NMA.}
\usage{
plotesthelper(x, lower, upper, estimate, xdata, xtitle, method)
}
\arguments{
\item{x}{object of class NMAoutlier or class NMAoutlier.rsv (mandatory).}

\item{lower}{lower bundary of confidence interval of summary estimate.}

\item{upper}{upper bundary of confidence interval of summary estimate.}

\item{estimate}{summary estimate.}

\item{xdata}{data.}

\item{xtitle}{title for plots in x-axis.}

\item{method}{The method that was used.
Select "fs" to monitor statistics during the forward search algorithm.
Select "rsv" to monitor statistics by fitting the data with the Shift Variance Model.}
}
\description{
Helper function to create the plots of summary estimates and their intervals for FS and RSV NMA.
}
\details{
Plot of summary estimate and its confidence interval for each treatment for FS and RSV NMA methodologies.
Vertical axis provides the iterations of FS methodology or the study for RSV NMA methodology.
Horizontal axis provides a summary estimate of a treatment.
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{internal}
