% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelper.R
\name{plothelper}
\alias{plothelper}
\title{Helper function to create the plots for FS and RSV NMA methodologies.}
\usage{
plothelper(x, method, stat, select.st)
}
\arguments{
\item{x}{object of class NMAoutlier or class NMAoutlier.rsv (mandatory).}

\item{method}{The method that was used.
Select "fs" to monitor statistics during the forward search algorithm.
Select "rsv" to monitor statistics by fitting the data with the Shift Variance Model.}

\item{stat}{The monitored statistic to generate the plot.}

\item{select.st}{selected statistic (pscore/nsplit/estand) for
selected treatment(s)/comparison(s)/study}
}
\description{
Helper function to create the plots for FS and RSV NMA methodologies.
}
\details{
Plot of several monitoring measures for FS and RSV NMA methodologies.
Vertical axis provides the iterations of FS methodology or the study for RSV NMA methodology.
Horizontal axis provides a monitoring statistical measure in the methodology.
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{internal}
