% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qnetplot.R
\name{Qnetplot}
\alias{Qnetplot}
\title{Q-Q plot for network meta-analysis (Q-Q netplot).}
\usage{
Qnetplot(data)
}
\arguments{
\item{data}{object of class NMAoutlier_measures (mandatory).}
}
\description{
This function generates the Q-Q plot (Q-Q net plot) for network meta-analyisis model.
}
\details{
Plot of Q-squared Mahalanobis distance for each study in the network.
Vertical axis provides the Q-squared Mahalanobis distance for each i study included in the network.
Horizontal axis provides the Q estimated quantile. Q estimated quantiles are the theoretical quantiles from the normal distribution.
A reference line is fitted from the cartesian points of the two measures. The normal Q-Q netplot tool allows to detect observation(s)
that are away from the reference line. This can be
an indication that normality cannot be assumed in the data and an outlying study exist.

Q-Q plot for network meta-analysis (Q-Q netplot) has been introduced in article by Petropoulou et al. 2019.
}
\examples{
data(smokingcessation, package = "netmeta")

p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm = "OR")

# outlier detection measures to compute Mahalanobis distance
measures <- NMAoutlier_measures(p1)

# Mahalanobis distance values for each study in the network
measures$Mah

# Q-Q netplot for the network of smoking cessation dataset
Qnetplot(measures)

}
\references{
Petropoulou M (2019):
Outlier detection measures in network meta-analysis.
\emph{Manuscript}
}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
\keyword{hplot}
