% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_multi.R
\name{res_multi}
\alias{res_multi}
\title{Statistical measure in multi-arm studies}
\usage{
res_multi(st.lab, ei)
}
\arguments{
\item{st.lab}{studylab of set of studies.}

\item{ei}{statistical measure for each pairwise comparison of studies.}
}
\value{
the arithmetic mean of statistical measure.
}
\description{
Computes the arithmetic mean of a statistical measure for the pairwise comparisons to study level data
(this can be done in the case of multi-arm trials).
}
\keyword{internal}
