% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvplot.R
\name{rsvplot}
\alias{rsvplot}
\title{Plot(s) to monitor selected statistic(s) with the
Random Shift Variance Network Meta-analysis (RSV NMA) model for selected/all studies.}
\usage{
rsvplot(x, stat, select.st = NULL)
}
\arguments{
\item{x}{an object of class NMAoutlier.rsv (mandatory).}

\item{stat}{statistical measure to be monitored
(mandatory), available choices are: "pscore", "estand",
"heterog",  "Q", "over_disp", "leverage",  "LRT" (can be abbreviated).}

\item{select.st}{selected statistic (pscore/estand) for
selected treatment(s)/comparison(s)/study}
}
\description{
The function creates plot(s) to monitor selected
statistic(s) and/or method(s) by shifting the variance for study (or for selected studies).
Candidate statistics to be monitored can be: P-score;
z-values by back-calculation method to derive indirect estimates
from direct pairwise comparisons and network estimates;
standardized residuals; heterogeneity variance estimator; over-dispersion parameter (shift variance estimator);
leverage; Q statistics (Overall heterogeneity /
inconsistency Q statistic (\code{Q}), overall heterogeneity Q
statistic (\code{Q}), between-designs Q statistic (\code{Q}), based
on a random effects design-by-treatment interaction model) and
likelihood ratio test (LRT).
}
\details{
Plot of monitoring measures by shifting the variance for each study (RSV NMA model).
Vertical axis provides the studies. Horizontal axis provides monitoring statistic(s) for RSV NMA model.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm="OR")

# RSV NMA model for study 1 of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad", study = c(1), n_cores = 2)

# Over dispresion for RSV NMA model fitted for study 1
rsvplot(RSVresult, "over_disp")

\dontrun{

# Random Shift Variance model for each study of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad")

# Heterogeneity variance estimator for RSV NMA model
rsvplot(RSVresult, "heterog")

# Over dispresion for RSV NMA model
rsvplot(RSVresult, "over_disp")

# Q statistics for RSV NMA model
rsvplot(RSVresult, "Q")

# P-scores for RSV NMA model
rsvplot(RSVresult, "pscore")

# Likelihood Ration Test (LRT) for RSV NMA model
rsvplot(RSVresult, "LRT")
}

}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{hplot}
