% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvplotest.R
\name{rsvplotest}
\alias{rsvplotest}
\title{Plots for summary treatment estimates in RSV NMA model.}
\usage{
rsvplotest(x)
}
\arguments{
\item{x}{object of class NMAoutlier.rsv (mandatory).}
}
\description{
Plot for summary treatment estimates for RSV NMA model.
}
\details{
Plot of summary treatment estimates and their confidence intervals
 RSV NMA model fitted for each study.
Vertical axis provides study. Horizontal axis
provides summary treatment estimates.
}
\examples{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data=smokingcessation,
                        sm="OR")

# RSV NMA model for study 1 of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad", study = c(1), n_cores = 2)

# Plot for summary treatment estimates
# and their confidence intervals with RSV NMA model
# fitted for study 1 (shift the variance of study 1)
rsvplotest(RSVresult)

\dontrun{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data=smokingcessation,
                        sm="OR")

# RSV NMA model for each study of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad")

# Plot for summary treatment estimates and their confidence intervals
# for RSV NMA model fitted for each study
rsvplotest(RSVresult)
}

}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{hplot}
