% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset.R
\name{Subset}
\alias{Subset}
\title{Subset}
\usage{
Subset(m, n, t1.label, t2.label, studlab, studies)
}
\arguments{
\item{m}{number of pairwise comparisons (edges)}

\item{n}{number of treatments (vertices)}

\item{t1.label}{numbers to treatment 1 IDs.}

\item{t2.label}{numbers to treatment 2 IDs.}

\item{studlab}{Study labels (important when multi arm studies are included).}

\item{studies}{An optional vector specifying a subset of studies to be used.
The default value is the number of studies.}
}
\value{
subset of studies that includes all treatments.
}
\description{
Finds a subset of studies with size equal to the initial set that includes all treatments.
}
