% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svrplotest.R
\name{svrplotest}
\alias{svrplotest}
\title{Plot for summary estimates for each treatment
and their confidence intervals
for Random Shift Variace Model
fitted for selected/all studies of a dataset.}
\usage{
svrplotest(x)
}
\arguments{
\item{x}{object of class NMAsvr (mandatory).}
}
\description{
Plot for summary estimates for each treatment
and their 95 percent confidence intervals for Random Shift Variace Model
fitted for selected or all studies of a dataset
}
\details{
Plot of summary estimates and their confidence intervals for each treatment
for Random Shift Variance Model fitted for each study.
Vertical axis provides study. Horizontal axis
provides summary estimate of a treatment.
}
\examples{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data=smokingcessation,
                        sm="OR")

# Random Shift Variace Model for study 1 of smoking cessation data
SVRresult <- NMAsvr(p1, small.values = "bad", study = c(1), n_cores = 2)

# Plot for summary estimates for each treatment
# and their confidence intervals for Random Shift Variace Model
# fitted for study 1
svrplotest(SVRresult)

\dontrun{
data(smokingcessation, package = "netmeta")

# Transform data from arm-based format to contrast-based format
# We use 'sm' argument for odds ratios.
# We use function pairwise from netmeta package
#
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data=smokingcessation,
                        sm="OR")

# Random Shift Variace Model for each study of smoking cessation data
SVRresult <- NMAsvr(p1, small.values = "bad")

# Plot for summary estimates for each treatment
# and their confidence intervals for Random Shift Variace Model
# fitted for each study
svrplotest(SVRresult)
}

}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
\keyword{hplot}
