% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdplot.R
\name{fwdplot}
\alias{fwdplot}
\title{Forward plot(s) to monitor selected statistic(s)/method(s).}
\usage{
fwdplot(x, stat, select.st = "NULL")
}
\arguments{
\item{x}{an object of class NMAoutlier (results from forward search algorithm in network meta-analysis).}

\item{stat}{measure(s) to be monitored in forward plot(s), available choices pscore, nsplit, estand, heterog, cook, ratio, Q.}

\item{select.st}{choice to monitor seleced statistic(s) (P-scores/node-splitting z-value/standardized resiluals) for selected item(s) (treatment(s)/comparison(s)/study(studies))}
}
\value{
forward plot(s) for selected statistic(s)/method(s).
}
\description{
Forward plot(s) to monitor selected statistic(s) and/or method(s) (P-score, Node-splitting, Standardized residuals, heterogeneity, cook distance, ratio of variances, Q statistics).
}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
