% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdplotest.R
\name{fwdplotest}
\alias{fwdplotest}
\title{Forward plots of summary estimates.}
\usage{
fwdplotest(x)
}
\arguments{
\item{x}{object of class NMAoutlier (results of FS algorithm).}
}
\value{
forward plots of summary estimates.
}
\description{
Forward plots of summary estimate with 95 percent confidence interval for each treatment.
}
\author{
Maria Petropoulou <mpetrop@cc.uoi.gr>
}
