% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAoutlier.rsv.R
\name{NMAoutlier.rsv}
\alias{NMAoutlier.rsv}
\title{Random Shift Variance Network Meta-analysis (RVS NMA) model.}
\usage{
NMAoutlier.rsv(
  TE,
  seTE,
  treat1,
  treat2,
  studlab,
  data = NULL,
  sm,
  reference = "",
  small.values = "good",
  study = NULL,
  n_cores = NULL
)
}
\arguments{
\item{TE}{Estimate of treatment effect, i.e. difference between
first and second treatment (e.g. log odds ratio, mean difference,
or log hazard ratio).}

\item{seTE}{Standard error of treatment estimate.}

\item{treat1}{Label/Number for first treatment.}

\item{treat2}{Label/Number for second treatment.}

\item{studlab}{Study labels (important when multi arm studies are
included).}

\item{data}{A data frame containing the study information.}

\item{sm}{A character string indicating underlying summary measure,
e.g., \code{"RD"}, \code{"RR"}, \code{"OR"}, \code{"ASD"},
\code{"HR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{reference}{Reference treatment group.}

\item{small.values}{A character string indicating if small values
are considered beneficial (option:"good") or harmfull
(option:"bad") on outcome.  This is requirement for p-scores
computation. The default value is considered benefial outcome
("good").}

\item{study}{The study (or studies) to fit by shift its (their) variance.
By default, the standard random effects model is fitted, value NULL.}

\item{n_cores}{The number of cores that the process is running
using the parallel (default: NULL, the process is running
using all the available cores)}
}
\value{
An object of class \code{NMAoutlier}; a list containing the
following components:
   \item{dat}{Matrix containing the data \code{"TE"}, \code{"seTE"}, \code{"studlab"}, \code{"treat1"}, \code{"treat2"} as defined above.}
   \item{length.initial}{The number of studies that constitute the initial (outlying-clean) subset of studies.}
   \item{index}{The number of iterations of forward search algorithm.}
   \item{basic}{Studies entered into the basic set in each iteration of the search.
   At the first iteration, basic set constitute the studies that are included in the basic-initial subset.
   The number of studies in the first iteration is equal to length.initial.}
   \item{taub}{Heterogeneity estimator variance for basic set in each iteration of forward search algorithm.}
   \item{Qb}{Overall heterogeneity - inconsistency Q statistic (\code{Q}) for the basic set in each iteration of forward search algorithm.}
   \item{Qhb}{Overall heterogeneity Q statistic (\code{Q}) for the basic set in each iteration of forward search algorithm.}
   \item{Qib}{Overall inconsistency Q statistic (\code{Q}) from design-by-treatment interaction model for the basic set in each iteration of forward search algorithm.}
   \item{estb}{Summary estimates for each treatment for the basic set in each iteration of forward search algorithm.}
   \item{lb}{Lower 95\% confidence interval of summary estimates for the basic set in each iteration of forward search algorithm.}
   \item{ub}{Upper 95\% confidence interval of summary estimates for the basic set in each iteration of forward search algorithm.}
   \item{Ratio}{Ratio of determinants (\code{COVRATIOj}) of variance-covariance matrix of treatment estimates at iteration j to that iteration at (j-1).}
   \item{cook_d}{Cook's statistic (\code{Cj}) at iteration j of forward search algorithm.}
   \item{p.score}{P-score for ranking each treatment for the basic set in each iteration of forward search algorithm.}
   \item{dif}{Z-values for comparison between direct and indirect evidence for each iteration of forward search algorithm.
    Based on back-calculation method to derive indirect estimates from direct pairwise comparisons and network estimates.}
   \item{estand}{Standardized residuals for each study for the basic set in each iteration of forward search algorithm.}
   \item{call}{Function call}
}
\description{
Fits the Random Shift Variance Network Meta-analysis Model for detection of outlying
studies (studies with extreme results). This is a diagnostic tool
for outlier detection. It can also be used to detect
studies that are potential sources for heterogeneity and
inconsistency.

Monitoring measures during the search are:
\itemize{
\item outlying measures (standardized residuals, leverage);
\item ranking measures (P-scores);
\item heterogeneity and inconsistency measures (Q statistics for
  overall heterogeneity / inconsistency, inconsistency by
  design-by-treatment interaction model, z-values for comparison
  between direct and indirect evidence by back-calculation method).
 \item over dispresion parameter
 \item Likelihood Ratio Test (LRT)
}

A description of the methodology can be found in Petropoulou et
al. (2020).
}
\details{
Fits the RVS NMA model for detecting outliers.


Description of methodology by fitting Random Shift Variance Network Meta-Analysis Model (RVS NMA) in
network meta-analysis. Methodology of RVS NMA model is described in Petropoulou (2020).


This function employs the Random Shift Variance Network Meta-analysis Model (RVS).
The RVS model is fitted by shifting the variance for each included study in the network.
RVS NMA model is fitted in network meta-analysis model
from graph theory (Rücker, 2012) (\code{netmeta} function) with
R package \bold{netmeta} (Rücker et al., 2015) using the REstricted Maximum Likelihood (REML)
estimation method for heterogeneity variance. The variance estimator of RVS NMA model is
decomposed by the heterogeneity variance and the over-dispersion.
The researcher can choose the reference treatment (\code{reference}) fitted in NMA model.
Monitoring is helpful to identify outlying and/or influential studies.

Monitoring statistical measures (outlying measures,
heterogeneity and inconsistency measures, ranking measures)
and likelihood statistics can be:


Monitoring is helpful to identify outlying studies:

\bold{Likelihood statistics:}
Heterogeneity estimation method is conducted under
the Restricted Maximum likelihood estimator.
The variance estimator of shift variance model.
Likelihood statistics offered from calculation:
the twice of maximum log-likelihood;
the convergence diagnostic and the Likelihood Ratio test (LRT) test.

\bold{Outlier and influential case diagnostics measures:}
Standardized residuals (arithmetic mean in case of multi-arm
studies); leverage of hat matrix

\bold{Ranking measures:}
P-scores for ranking of treatments (Rücker G & Schwarzer G (2015))
for each basic set with implementation of (\code{netrank} function)
from R package \bold{netmeta}.

\bold{Heterogeneity and inconsistency measures:}
Overall heterogeneity / inconsistency Q statistic (\code{Q}) This
is the design-based decomposition of Cochran Q as provided by Krahn
et al.(2013); Overall heterogeneity Q statistic (\code{Q});
Between-designs Q statistic (\code{Q}), based on a random effects
model with square-root of between-study variance estimated embedded
in a full design-by-treatment interaction model. Implementation
with (\code{decomp.design} function) from R package \bold{netmeta};
Z-values for comparison between direct and indirect evidence for
each iteration of forward search algorithm.  By monitoring
difference of direct and indirect evidence, potential sources of
consistency can be detected with the implementation of
(\code{netsplit} function) from R package \bold{netmeta} for each
basic set of the search.  Based on the methodology with
back-calculation method to derive indirect estimates from direct
pairwise comparisons and network estimates (Dias et al., 2010;
König et al., 2013).
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm="OR")

# Random Shift Variance NMA model implementation for study 1
#
RVSOMres <- NMAoutlier.rsv(p1,  small.values = "bad", study = c(1) , n_cores = 2)
#

\dontrun{
# Random Shift Variance Model
#
RVSOMresult <- NMAoutlier.rsv(p1, small.values = "bad")

# Random shift variance estimator
RVSOMresult$over-disp
#
# Plot of Likelihood Ratio Test
RVSOMresult$LRT
#
# Summary estimators and their 95\% confidence intervals for each treatment
# by downweighting each study
RVSOMresult$b
RVSOMresult$l
RVSOMresult$u
}

}
\references{
Rücker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24

Rücker G, Schwarzer G (2015):
Ranking treatments in frequentist network meta-analysis works
without resampling methods.
\emph{BMC Medical Research Methodology},
\bold{15}, 58

Petropoulou M (2020):
Exploring methodological challenges in network meta-analysis models and
developing methodology for outlier detection.
\emph{PhD dissertation}.

Krahn U, Binder H, König J (2013):
A graphical tool for locating inconsistency in network meta-analyses.
\emph{BMC Medical Research Methodology},
\bold{13}, 35

Dias S, Welton NJ, Caldwell DM, Ades AE (2010):
Checking consistency in mixed treatment comparison meta-analysis.
\emph{Statistics in Medicine},
\bold{29}, 932--44

König J, Krahn U, Binder H (2013):
Visualizing the flow of evidence in network meta-analysis and
characterizing mixed treatment comparisons.
\emph{Statistics in Medicine},
\bold{32}, 5414--29
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
