% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotesthelper.R
\name{plotesthelper}
\alias{plotesthelper}
\title{Helper function to create the plots of summary estimates and their intervals for FS algorithm.}
\usage{
plotesthelper(x, lower, upper, estimate, xdata, xtitle)
}
\arguments{
\item{x}{object of class NMAoutlier(mandatory).}

\item{lower}{lower bundary of confidence interval of summary estimate.}

\item{upper}{upper bundary of confidence interval of summary estimate.}

\item{estimate}{summary estimate.}

\item{xdata}{data.}

\item{xtitle}{title for plots in x-axis.}
}
\description{
Helper function to create the plots of summary estimates and their intervals for FS algorithm.
}
\details{
Plot of summary estimate and its confidence interval for each treatment for FS and algorithm.
Vertical axis provides the iterations of FS algorithm.
Horizontal axis provides a summary estimate of a treatment.
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{internal}
