% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelper.R
\name{plothelper}
\alias{plothelper}
\title{Helper function to create the plots for FS algorithm.}
\usage{
plothelper(x, stat, select.st)
}
\arguments{
\item{x}{object of class NMAoutlier (mandatory).}

\item{stat}{The monitored statistic to generate the plot.}

\item{select.st}{selected statistic (pscore/nsplit/estand) for
selected treatment(s)/comparison(s)/study}
}
\description{
Helper function to create the plots for FS algorithm.
}
\details{
Plot of several monitoring measures for FS algorithm.
Vertical axis provides the iterations of FS algorithm.
Horizontal axis provides a monitoring statistical measure in the methodology.
}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{internal}
