% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvplot.R
\name{rsvplot}
\alias{rsvplot}
\title{Plot(s) to monitor selected statistic(s)/method(s) for
the Random Shift Variance Network Meta-analysis (RSV NMA) model fitted for selected/all studies.}
\usage{
rsvplot(x, stat, select.st = NULL)
}
\arguments{
\item{x}{an object of class NMAoutlier.rsv (mandatory).}

\item{stat}{statistical measure to be monitored in forward plot(s)
(mandatory), available choice are: "pscore", "nsplit", "estand",
"heterog",  "Q", "over_disp", "leverage" or "LRT" (can be abbreviated).}

\item{select.st}{selected statistic (pscore/nsplit/estand) for
selected treatment(s)/comparison(s)/study}
}
\description{
The function creates a plot of a monitoring statistic of RSV NMA model which fitted for selected/all studies.
Candidate statistics to be monitored can be P-score;
z-values by back-calculation method to derive indirect estimates
from direct pairwise comparisons and network estimates;
standardized residuals; heterogeneity variance estimator; over-dispersion parameter;
leverage; Q statistics (Overall heterogeneity /
inconsistency Q statistic (\code{Q}), overall heterogeneity Q
statistic (\code{Q}), between-designs Q statistic (\code{Q}), based
on a random effects design-by-treatment interaction model) and
likelihood ratio test (LRT).
}
\details{
Plot of statistical measures for RSV NMA model fitted for each study.
Vertical axis provides studies. Horizontal axis provides monitoring statistic(s) of RSV NMA model which fitted for selected/all studies.
}
\examples{
data(smokingcessation, package = "netmeta")
smokingcessation$id <- 1:nrow(smokingcessation)
p1 <- netmeta::pairwise(list(treat1, treat2, treat3),
                        list(event1, event2, event3),
                        list(n1, n2, n3),
                        data = smokingcessation,
                        sm="OR")

# Random Shift Variance model for study 1 of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad", study = c(1), n_cores = 2)

# Over dispresion for Random Shift Variance model fitted for study 1
rsvplot(RSVresult, "over_disp")

\dontrun{

# Random Shift Variance model for each study of smoking cessation data
RSVresult <- NMAoutlier.rsv(p1, small.values = "bad")

# Heterogeneity variance estimator for Random Shift Variance model
rsvplot(RSVresult, "heterog")

# Over dispresion for Random Shift Variance model
rsvplot(RSVresult, "over_disp")

# Q statistics for Random Shift Variance model
rsvplot(RSVresult, "Q")

# P-scores for Random Shift Variance model
rsvplot(RSVresult, "pscore")

# z-values of disagreement of direct and indirect evidence for Random Shift Variance model
rsvplot(RSVresult, "nsplit")

# z-values of disagreement of direct and indirect evidence
# monitoring treatment comparison A versus B for Random Shift Variance model
rsvplot(RSVresult, "nsplit", "A:B")

# log likelihood test plot (LRT)
rsvplot(RSVresult, "LRT")
}

}
\author{
Maria Petropoulou <petropoulou@imbi.uni-freiburg.de>
}
\keyword{hplot}
